backend/requirements.txt
aiomysql==0.1.1
anyio==3.7.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
databases==0.7.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pydantic==1.10.8
PyMySQL==1.0.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
